// client/stocklinx app/src/redux/consumable/constant.ts
export enum consumableConst {
  FETCH_CONSUMABLES_REQUEST = "FETCH_CONSUMABLES_REQUEST",
  FETCH_CONSUMABLES_SUCCESS = "FETCH_CONSUMABLES_SUCCESS",
  FETCH_CONSUMABLES_FAILURE = "FETCH_CONSUMABLES_FAILURE",
  FETCH_CONSUMABLE_REQUEST = "FETCH_CONSUMABLE_REQUEST",
  FETCH_CONSUMABLE_SUCCESS = "FETCH_CONSUMABLE_SUCCESS",
  FETCH_CONSUMABLE_FAILURE = "FETCH_CONSUMABLE_FAILURE",
  CREATE_CONSUMABLE_REQUEST = "CREATE_CONSUMABLE_REQUEST",
  CREATE_CONSUMABLE_SUCCESS = "CREATE_CONSUMABLE_SUCCESS",
  CREATE_CONSUMABLE_FAILURE = "CREATE_CONSUMABLE_FAILURE",
  UPDATE_CONSUMABLE_REQUEST = "UPDATE_CONSUMABLE_REQUEST",
  UPDATE_CONSUMABLE_SUCCESS = "UPDATE_CONSUMABLE_SUCCESS",
  UPDATE_CONSUMABLE_FAILURE = "UPDATE_CONSUMABLE_FAILURE",
  REMOVE_CONSUMABLE_REQUEST = "REMOVE_CONSUMABLE_REQUEST",
  REMOVE_CONSUMABLE_SUCCESS = "REMOVE_CONSUMABLE_SUCCESS",
  REMOVE_CONSUMABLE_FAILURE = "REMOVE_CONSUMABLE_FAILURE",
  SET_CONSUMABLE = "SET_CONSUMABLE",
  CLEAR_CONSUMABLE = "CLEAR_CONSUMABLE",
  SET_CONSUMABLES = "SET_CONSUMABLES",
  CLEAR_CONSUMABLES = "CLEAR_CONSUMABLES",
}